refactor: ratchet diagnostics page size
