+++ frontend/dashboard/src/design/index.ts
@@ -1,6 +1,6 @@
 import './tokens.css';
 
-export { PageLoadProgress, type PageLoadProgressProps } from './PageLoadProgress';
+export { PageLoadProgress } from './PageLoadProgress';
 
 export {
   Button,
